requirements.txt
streamlit
mediapipe
Pillow
opencv-python-headless
numpy